Add base `upgrade` script
It's been painful working on my dotfiles on one machine, then piecing
together the "upgrade" path on another...so I've written a fairly basic
"upgrade" script, instead of having to comment out sections of the
"install" script.

So far, this is just an example...but I'll start actually writing
upgrade steps from here on out!