Add `jq` package to install scripts
The `jq` package will help with some parsing in a tmux weather plugin
I'm writing, so I've added it to the Brewfile to keep up.

I've also taken the opportunity to create an `apt` script that'll take
care of all the apt packages on Ubuntu, since the Homebrew-for-Linux
experiment doesn't seem to be going so well on my Raspberry Pi. I'll
keep moving packages over to this file in the future.