Ensure Tmux is installed on Ubuntu
When I installed on this Ubuntu box yesterday, I was confused when tpm
threw a bunch of errors, and then none of my normal plugins and
customizations were working.

Today, I realized it was probably because I hadn't actually installed
the latest version of tmux prior to installing tpm and its plugins,
which means nothing was loading properly!