Update Ubuntu Homebrew installation
When I originally installed this on my Ubuntu DO Droplet, it installed
Homebrew into my homedir (`~/.linuxbrew`), instead of a common location
(`/home/linuxbrew/.linuxbrew`)...meaning when I tried to add new stuff,
it would complain, and sometimes just not work.

I uninstalled/reinstalled to the correct location, and took the
opportunity to update the scripts accordingly as well.