Enable true color support in tmux
Once I fired up ghostty and connected to a tmux session, I noticed the
neovim colorscheme wasn't working properly...but it did work in a
non-tmux session. Apparently iTerm handled that for me?

Regardless, I found [a gist]() with suggested config to get the colors
working properly. According to an [FAQ]() in the tmux README, this is
because tmux needs to know that the outside terminal supports RGB color.

[a gist](https://gist.github.com/andersevenrud/015e61af2fd264371032763d4ed965b6)
[FAQ](https://github.com/tmux/tmux/wiki/FAQ#how-do-i-use-rgb-colour)