Add `lazygit` config
I kept getting random requests for my 1Password SSH key when I was doing
other things...traced it down to a tmux session, and realized I'd left
`lazygit` running - and it was trying to auto-fetch in the background.
Don't like that!

Took the opportunity to create a custom config file.