Configure `treesitter` plugin
The big learning here was to use the `master` branch, rather than
`main`. Somewhat confusingly, the `master` branch is an archived,
locked, backwards-compatible version; while the `main` branch "is a
full, incompatible, rewrite." Based on the amount of open issues and
open PRs, it seems it's not quite stable...and has a very tedious
migration process.

I don't care enough to try to make it happen...once I figured out how to
go back to the way things were, that's seems good enough for me.

(And now I wonder if this change would have just fixed the issues I was
running into with LazyVim to begin with...)